requirements.txt
# Graph Database
# real_ladybug>=0.1.0  # Not yet available
kuzu>=0.11.0  # Using Kuzu (same API as LadybugDB)

# SQL Database (built-in with Python)
# sqlite3 is included in Python standard library

# Utilities
python-dateutil>=2.8.0
mcp[cli]>=1.26.0
uvicorn>=0.24.0
starlette>=0.35.0

# Timestamp proofs (optional but recommended)
opentimestamps-client>=0.7.0

# Encrypted backups
pyzipper>=0.3.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
cryptography>=41.0.0
